requirements.txt
RPi.GPIO
spidev
Pillow
Flask
numpy
python-mpv
gunicorn
smbus
sugarpie
websockets
bech32

